Moving cpc modules -> curves
